Global Ecosystem State Variables

Global indicators of the ecosystem state, including richness, evenness, diversity, and other indicators

alpha(x, index = "all", zeroes = TRUE)

Arguments

x

A species abundance vector, or matrix (taxa/features x samples) with the absolute count data (no relative abundances), or phyloseq-class object

index

Default is ‘NULL’, meaning that all available indices will be included. For specific options, see details.

zeroes

Include zero counts in the diversity estimation.

Value

A data.frame of samples x alpha diversity indicators

Details

This function returns various indices of the ecosystem state. The function is named alpha (global in some previous versions of this package) as these indices can be viewed as measures of alpha diversity. The function uses default choices for detection, prevalence and other parameters for simplicity and standardization. See the individual functions for more options. All indicators from the richness, diversity, evenness, dominance, and rarity functions are available. Some additional measures, such as Chao1 and ACE are available via estimate_richness function in the phyloseq package but not included here. The index names are given the prefix richness_, evenness_, diversity_, dominance_, or rarity_ in the output table to avoid confusion between similarly named but different indices (e.g. Simpson diversity and Simpson dominance). All parameters are set to their default. To experiment with different parameterizations, see the more specific index functions (richness, diversity, evenness, dominance, rarity).

References

See citation('microbiome')

See also

Author

Contact: Leo Lahti microbiome-admin@googlegroups.com

Examples

data(dietswap)
d <- alpha(dietswap, index='shannon')
#> Observed richness
#> Other forms of richness
#> Diversity
#> Evenness
#> Dominance
#> Rarity
# d <- alpha(dietswap, index='all')