Core Microbiota

Filter the phyloseq object to include only prevalent taxa.

core(x, detection, prevalence, method = "standard", Nsample = NULL,
  bs.iter = 1000, I.max = NULL)

Arguments

x

phyloseq-class object

detection

Detection threshold (non-negative real)

prevalence

Prevalence threshold (in [0, 1])

method

Either "standard" or "bootstrap". The standard methods selects the taxa that exceed the given detection and prevalence threshold. The bootstrap method is more robust an described in Salonen et al. (2012). Note that the results may depend on the random seed unless a sufficiently large bootstrap sample size is used.

Nsample

Only needed for method "bootstrap". Bootstrap sample size, default is the same size as data.

bs.iter

Only needed for method "bootstrap". Bootstrap iterations.

I.max

Only needed for method "bootstrap". Upper limit for intensity threshold. Later addition. Set to NULL (default) to replicate Salonen et al.

...

Arguments to pass.

Value

Filtered phyloseq object including only prevalent taxa

References

The core microbiota bootstrap method implemented with this function: Salonen A, Salojarvi J, Lahti L, de Vos WM. The adult intestinal core microbiota is determined by analysis depth and health status. Clinical Microbiology and Infection 18(S4):16-20, 2012 To cite the microbiome R package, see citation('microbiome')

Examples

  data(peerj32)
  core(peerj32$phyloseq, 200, .2)

#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 42 taxa and 44 samples ]
#> sample_data() Sample Data:       [ 44 samples by 5 sample variables ]
#> tax_table()   Taxonomy Table:    [ 42 taxa by 2 taxonomic ranks ]