Bootstrap Analysis of the Core Microbiota

Bootstrap analysis of the core microbiota.

core_bootstrap(x, Nsample = NULL, prevalence = 0.5, bs.iter = 1000,
  detection = 0, I.max = NULL)

Arguments

x

OTUxSample data matrix

Nsample

bootstrap sample size, default is the same size as data

prevalence

Lower limit for number of samples where microbe needs to exceed the intensity threshold for a 'present' call.

bs.iter

bootstrap iterations

detection

Lower limit for intensity threshold

I.max

Upper limit for intensity threshold. Later addition. set to NULL (default) to replicate Salonen et al.

Value

data frame with microbes and their frequency of presence in the core

References

The core microbiota bootstrap method implemented with this function: Salonen A, Salojarvi J, Lahti L, de Vos WM. The adult intestinal core microbiota is determined by analysis depth and health status. Clinical Microbiology and Infection 18(S4):16-20, 2012 To cite this R package, see citation("microbiome")

Examples

  data(peerj32)
  # In practice, use bs.iter = 1000 or more
  bs <- core(peerj32$phyloseq, bs.iter = 5,
             prevalence = .5, detection = 0,
        method = "bootstrap")

#> Bootstrapped median core size: 18

  # Not exported:
  # bs <- core_bootstrap(peerj32$phyloseq, bs.iter = 5)