Core Heatmap

Core heatmap.

core_heatmap(x, dets, cols, min.prev, taxa.order)

Arguments

x

OTU matrix

dets

A vector or a scalar indicating the number of intervals in (0, log10(max(data))). The dets are calculated for relative abundancies.

cols

colours for the heatmap

min.prev

If minimum prevalence is set, then filter out those rows (taxa) and columns (dets) that never exceed this prevalence. This helps to zoom in on the actual core region of the heatmap.

taxa.order

Ordering of the taxa.

Value

Used for its side effects

References

A Salonen et al. The adult intestinal core microbiota is determined by analysis depth and health status. Clinical Microbiology and Infection 18(S4):16 20, 2012. To cite the microbiome R package, see citation('microbiome')

Author

Contact: Leo Lahti microbiome-admin@googlegroups.com