Estimate Diversity

Diversity estimation. Augments the estimate_richness function of the phyloseq package.

diversity(x, detection = 0, split = TRUE, measures = NULL)

Arguments

x

phyloseq-class object

detection

detection for observing taxa (absence / presence). Used to determine the richness (Observed diversity) above this abundance threshold. Zero by default.

split

(Optional). Logical. Should a separate set of richness estimates be performed for each sample? Or alternatively, pool all samples and estimate richness of the entire set.

measures

(Optional). Default is ‘NULL’, meaning that all available alpha-diversity measures will be included. Alternatively, you can specify one or more measures as a character vector of measure names. Values must be among those supported in the phyloseq::estimate_richness function. These include ‘c("Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher")’. In addition, the measure "Evenness" is provided (Pielou's index).

Value

A data.frame of samples x diversity indicators; except when split=FALSE, a vector of indices is returned.

References

See citation('microbiome')

Examples

  data(dietswap)
  d <- diversity(dietswap)