Get Ordination

Ordinate phyloseq data and merge it with sample metadata

get_ordination(x, method = "NMDS", distance = "bray")

Arguments

x

phyloseq-class object or a data matrix (features x samples; eg. HITChip taxa vs. samples)

method

Ordination method, see phyloseq::plot_ordination

distance

Ordination distance, see phyloseq::plot_ordination

Value

data.frame with ordination coordinates and metadata

Details

This is a wrapper for phyloseq ordination functions, providing smooth access to ordinated data.frame with full info on the projection and metadata necessary for further visualizations.

References

See citation('microbiome')

See also

phyloseq::plot_ordination

Examples

  data(dietswap)
  fc <- get_ordination(dietswap)

#> Square root transformation
#> Wisconsin double standardization
#> Run 0 stress 0.2044329 
#> Run 1 stress 0.2061352 
#> Run 2 stress 0.2095027 
#> Run 3 stress 0.2145099 
#> Run 4 stress 0.2036035 
#> ... New best solution
#> ... Procrustes: rmse 0.01111379  max resid 0.1378948 
#> Run 5 stress 0.4170742 
#> Run 6 stress 0.2049269 
#> Run 7 stress 0.2117836 
#> Run 8 stress 0.2059434 
#> Run 9 stress 0.2160625 
#> Run 10 stress 0.2092597 
#> Run 11 stress 0.2037527 
#> ... Procrustes: rmse 0.005896076  max resid 0.08546946 
#> Run 12 stress 0.2047395 
#> Run 13 stress 0.2129574 
#> Run 14 stress 0.20986 
#> Run 15 stress 0.2056086 
#> Run 16 stress 0.212241 
#> Run 17 stress 0.2136399 
#> Run 18 stress 0.2119387 
#> Run 19 stress 0.2042779 
#> Run 20 stress 0.2059905 
#> *** No convergence -- monoMDS stopping criteria:
#>     20: stress ratio > sratmax