Association Heatmap

Visualizes n x m association table as heatmap.

heat(df, Xvar, Yvar, fill, star, p.adj.threshold = 1,
  association.threshold = 0, step = 0.2, colours = c("darkblue", "blue",
  "white", "red", "darkred"), limits = NULL, legend.text = "",
  order.rows = TRUE, order.cols = TRUE, text.size = 10,
  filter.significant = TRUE, star.size = NULL, plot.values = FALSE)

Arguments

df

Data frame. Each row corresponds to a pair of associated variables. The columns give variable names, association scores and significance estimates.

Xvar

X axis variable column name. For instance 'X'.

Yvar

Y axis variable column name. For instance 'Y'.

fill

Column to be used for heatmap coloring. For instance 'association'.

star

Column to be used for cell highlighting. For instance 'p.adj'.

p.adj.threshold

Significance threshold for the stars.

association.threshold

Include only elements that have absolute association higher than this value

step

color interval

colours

heatmap colours

limits

colour scale limits

legend.text

legend text

order.rows

Order rows to enhance visualization interpretability

order.cols

Order columns to enhance visualization interpretability

text.size

Adjust text size

filter.significant

Keep only the elements with at least one significant entry

star.size

NULL Determine size of the highlight symbols

plot.values

Show values as text

Value

ggplot2 object

References

See citation('microbiome')

Examples

  data(peerj32)
  d1 <- peerj32$lipids[, 1:10]
  d2 <- peerj32$microbes[, 1:10]
  cc <- associate(d1, d2, method = "pearson")
  p <- heat(cc, 'X1', 'X2', 'Correlation', star = "p.adj")

#> Ordering columns

#> Ordering rows