Plot Diversity

Plot alpha diversity. This function estimates a number of alpha-diversity metrics using the estimate_richness function, and returns a ggplot object. The plot generated by this function will include every sample in physeq, but they can be further grouped on the horizontal axis through the argument to x, and shaded according to the argument to color (see below). You must use untrimmed, non-normalized count data for meaningful results.

plot_diversity(x, variable = "group", measures = "Shannon", nrow = 1,
  scales = "free_y", detection = 0, indicate.subjects = FALSE,
  na.rm = FALSE)

Arguments

x

phyloseq-class object

variable

A variable to map to the horizontal axis. The vertical axis will be mapped to the alpha diversity index/estimate and have units of total taxa, and/or index value (dimensionless). This parameter (x) is a character string indicating a in the dataset (nsamples(x)).

measures

Default is NULL. In this case all available alpha-diversity measures will be included. Alternatively, you can specify one or more measures as a character vector. Values must be among those supported: c("Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher").

nrow

Number of rows for plot faceting.

scales

scales for the plot

detection

Detection threshold for the diversity measure 'Observed' (ie. species richness). See diversity

indicate.subjects

Indicate subjects by lines. The sample_data(x) must have 'subject' field.

na.rm

Remove samples with missing metadata (NA)

Value

A ggplot plot object summarizing the richness estimates, and their standard error.

Details

If subject is among the metadata variables, the matched subjects across groups are indicated by lines.

See also

estimate_richness diversity plot_richness estimateR diversity

Examples

  # Load gut microbiota data on 1006 western adults
  # (see help(atlas1006) for references and details)
  data(atlas1006)
  # Visualize Shannon diversity across bmi groups; remove cases with no bmi info
  p <- plot_diversity(atlas1006, variable = "bmi_group", "Shannon", na.rm = TRUE)