Plot Potential

Visualization of the potential function.

plot_potential(res, cutoff = 0.5, plot.contours = TRUE, binwidth = 0.2,
  bins = NULL)

Arguments

res

output from potential_slidingaverage function

cutoff

parameter determining the upper limit of potential for visualizations

plot.contours

Plot contour lines.

binwidth

binwidth for contour plot

bins

bins for contour plot. Overrides binwidth if given

Value

A ggplot2 visualization of the potential landscape.

Details

Applied on the output of the potential_slidingaverage function.

References

Lahti et al. Tipping elements of the human intestinal ecosystem. Nature Communications 5:4344, 2014.

See also

potential_analysis

Examples

  ## Not run: ------------------------------------
# 
#   # This exemplifies visualization of data from an external package.
# 
#   # Load gut microbiota data on 1006 western adults
#   # From http://doi.org/10.5061/dryad.pk75d
#   # (see help(atlas1006) for references and details)
#   data(atlas1006)
#   pseq <- atlas1006
# 
#   # Pick diversity and age
#   diversity <- exp(microbiome::diversity(pseq)$Shannon)
#   age <- meta(pseq)$age
#   
#   # Run potential analysis
#   library(earlywarnings)
#   library(ggplot2)
#   res <- movpotential_ews(diversity, age)
#   plot_potential(res, cutoff = 0.5)
# 
#   
## ---------------------------------------------