Visually Weighted Regression Plot

Draw regression curve with smoothed error bars with Visually-Weighted Regression by Solomon M. Hsiang; see http://www.fight-entropy.com/2012/07/visually-weighted-regression.html The R is modified from Felix Schonbrodt's original code at http://www.nicebread.de/visually-weighted-watercolor-plots-new-variants-please-vote

plot_regression(formula, data, B = 1000, shade = TRUE, shade.alpha = 0.1,
  spag = FALSE, mweight = TRUE, show.lm = FALSE, show.median = TRUE,
  median.col = "white", show.CI = FALSE, method = loess, bw = FALSE,
  slices = 200, palette = colorRampPalette(c("#FFEDA0", "#DD0000"), bias =
  2)(20), ylim = NULL, quantize = "continuous", show.points = TRUE, ...)

Arguments

formula

formula

data

data

B

number bootstrapped smoothers

shade

plot the shaded confidence region?

shade.alpha

shade.alpha: should the CI shading fade out at the edges? (by reducing alpha; 0 = no alpha decrease, 0.1 = medium alpha decrease, 0.5 = strong alpha decrease)

spag

plot spaghetti lines?

mweight

should the median smoother be visually weighted?

show.lm

should the linear regresison line be plotted?

show.median

show median smoother

median.col

median color

show.CI

should the 95% CI limits be plotted?

method

the fitting function for the spaghettis; default: loess

bw

define a default b/w-palette (TRUE/FALSE)

slices

number of slices in x and y direction for the shaded region. Higher numbers make a smoother plot, but takes longer to draw. I wouldn'T go beyond 500

palette

provide a custom color palette for the watercolors

ylim

restrict range of the watercoloring

quantize

either "continuous", or "SD". In the latter case, we get three color regions for 1, 2, and 3 SD (an idea of John Mashey)

show.points

Show points.

...

further parameters passed to the fitting function, in the case of loess, for example, "span = .9", or "family = 'symmetric'"

Value

ggplot2 object

References

See citation("microbiome")

Examples

  ## Not run: ------------------------------------
#     data(atlas1006)
#     p <- plot_regression(diversity ~ age, sample_data(atlas1006))
#   
## ---------------------------------------------