Bootstrapped Potential Analysis

Analysis of multimodality based on bootstrapped potential analysis of Livina et al. (2010) as described in Lahti et al. (2014).

potential_analysis(x, peak.threshold = 0, bw.adjust = 1, bs.iter = 100,
  min.density = 1)

Arguments

x

Input data vector

peak.threshold

Mode detection threshold

bw.adjust

Bandwidth adjustment

bs.iter

Bootstrap iterations

min.density

minimum accepted density for a maximum; as a multiple of kernel height

Value

List with following elements:

modesNumber of modes for the input data vector (the most frequent number of modes from bootstrap)

minimaAverage of potential minima across the bootstrap samples (for the most frequent number of modes)

maximaAverage of potential maxima across the bootstrap samples (for the most frequent number of modes)

unimodality.supportFraction of bootstrap samples exhibiting unimodality

bwsBandwidths

References

Livina et al. (2010). Potential analysis reveals changing number of climate states during the last 60 kyr. Climate of the Past, 6, 77-82.

Lahti et al. (2014). Tipping elements of the human intestinal ecosystem. Nature Communications 5:4344.

See also

plot_potential

Examples

# Example data; see help(peerj32) for details
data(peerj32)

# Log10 abundance of Dialister
x <- log10(abundances(peerj32$phyloseq)["Dialister",])

# Bootstrapped potential analysis
res <- potential_analysis(x, peak.threshold = 0, bw.adjust = 1, bs.iter = 100, min.density = 1)