Prevalence for Phyloseq OTUs

Simple prevalence measure.

prevalence(x, detection = 0, sort = FALSE, count = FALSE)

Arguments

x

A vector, data matrix or phyloseq object

detection

Detection threshold for absence/presence.

sort

Sort the groups by prevalence

count

Logical. Indicate prevalence as fraction of samples (in percentage [0, 1]; default); or in absolute counts indicating the number of samples where the OTU is detected above the given abundance threshold.

Value

For each OTU, the fraction of samples where a given OTU is detected. The output is readily given as a percentage.

Details

For vectors, calculates the fraction (count = FALSE) or number (count = TRUE) of samples that exceed the detection. For matrices, calculates this for each matrix column. For phyloseq objects, calculates this for each OTU. The relative prevalence (count = FALSE) is simply the absolute prevalence (count = TRUE) divided by the number of samples.

References

A Salonen et al. The adult intestinal core microbiota is determined by analysis depth and health status. Clinical Microbiology and Infection 18(S4):16 20, 2012. To cite the microbiome R package, see citation('microbiome')

Examples

  data(peerj32)
  ## With matrix
  prevalence(peerj32$data$microbes, detection = 200, sort = TRUE)

#> Warning: x is NULL - returning NULL

#> NULL

  ## With phyloseq
  prevalence(peerj32$phyloseq, detection = 200, sort = TRUE)

#>            Ruminococcus obeum et rel.        Roseburia intestinalis et rel. 
#>                            1.00000000                            1.00000000 
#>  Faecalibacterium prausnitzii et rel.           Eubacterium rectale et rel. 
#>                            1.00000000                            1.00000000 
#>            Eubacterium hallii et rel.         Dorea formicigenerans et rel. 
#>                            1.00000000                            1.00000000 
#>          Coprococcus eutactus et rel.            Clostridium nexile et rel. 
#>                            1.00000000                            1.00000000 
#>           Anaerostipes caccae et rel.         Ruminococcus lactaris et rel. 
#>                            1.00000000                            0.97727273 
#>         Clostridium symbiosum et rel.        Clostridium sphenoides et rel. 
#>                            0.97727273                            0.95454545 
#>      Bryantella formatexigens et rel.      Subdoligranulum variable at rel. 
#>                            0.95454545                            0.93181818 
#>   Papillibacter cinnamivorans et rel.        Eubacterium ventriosum et rel. 
#>                            0.93181818                            0.90909091 
#>          Bacteroides vulgatus et rel.     Lachnospira pectinoschiza et rel. 
#>                            0.90909091                            0.88636364 
#>          Lachnobacillus bovis et rel.                       Bifidobacterium 
#>                            0.88636364                            0.88636364 
#>           Streptococcus bovis et rel.           Ruminococcus gnavus et rel. 
#>                            0.86363636                            0.86363636 
#>           Ruminococcus bromii et rel.                           Akkermansia 
#>                            0.77272727                            0.77272727 
#>           Streptococcus mitis et rel.         Bacteroides uniformis et rel. 
#>                            0.75000000                            0.72727273 
#>      Bacteroides intestinalis et rel.  Outgrouping clostridium cluster XIVa 
#>                            0.70454545                            0.61363636 
#>            Clostridium leptum et rel.    Oscillospira guillermondii et rel. 
#>                            0.59090909                            0.56818182 
#>      Clostridium orbiscindens et rel.                    Allistipes et rel. 
#>                            0.54545455                            0.50000000 
#>          Bacteroides fragilis et rel.        Oxalobacter formigenes et rel. 
#>                            0.38636364                            0.36363636 
#>    Parabacteroides distasonis et rel.         Clostridium cellulosi et rel. 
#>                            0.34090909                            0.34090909 
#>        Butyrivibrio crossotus et rel.         Ruminococcus callidus et rel. 
#>                            0.31818182                            0.29545455 
#>        Sporobacter termitidis et rel.           Prevotella tannerae et rel. 
#>                            0.27272727                            0.25000000 
#>         Bacteroides stercoris et rel.            Bacteroides ovatus et rel. 
#>                            0.20454545                            0.20454545 
#>            Uncultured Clostridiales I          Bacteroides plebeius et rel. 
#>                            0.18181818                            0.18181818 
#>           Eubacterium biforme et rel.                             Dialister 
#>                            0.15909091                            0.15909091 
#>     Prevotella melaninogenica et rel.              Uncultured Bacteroidetes 
#>                            0.13636364                            0.11363636 
#>     Streptococcus intermedius et rel.                           Collinsella 
#>                            0.09090909                            0.09090909 
#>       Anaerovorax odorimutans et rel.             Prevotella oralis et rel. 
#>                            0.09090909                            0.06818182 
#>                           Lactococcus                              Serratia 
#>                            0.06818182                            0.04545455 
#>     Catenibacterium mitsuokai et rel.              Uncultured Chroococcales 
#>                            0.04545455                            0.02272727 
#>                           Pseudomonas         Prevotella ruminicola et rel. 
#>                            0.02272727                            0.02272727 
#>        Mitsuokella multiacida et rel.      Eubacterium cylindroides et rel. 
#>                            0.02272727                            0.02272727 
#>              Escherichia coli et rel.         Clostridium difficile et rel. 
#>                            0.02272727                            0.02272727 
#>           Clostridium colinum et rel.       Bacteroides splachnicus et rel. 
#>                            0.02272727                            0.02272727 
#>                         Aquabacterium                      Yersinia et rel. 
#>                            0.02272727                            0.00000000 
#>                      Xanthomonadaceae                      Wissella et rel. 
#>                            0.00000000                            0.00000000 
#>                                Vibrio                     Weissella et rel. 
#>                            0.00000000                            0.00000000 
#>                           Veillonella           Uncultured Selenomonadaceae 
#>                            0.00000000                            0.00000000 
#>                 Uncultured Mollicutes           Uncultured Clostridiales II 
#>                            0.00000000                            0.00000000 
#>                    Tannerella et rel.        Sutterella wadsworthia et rel. 
#>                            0.00000000                            0.00000000 
#>                        Staphylococcus                       Proteus et rel. 
#>                            0.00000000                            0.00000000 
#>                     Propionibacterium Phascolarctobacterium faecium et rel. 
#>                            0.00000000                            0.00000000 
#>     Peptostreptococcus micros et rel. Peptostreptococcus anaerobius et rel. 
#>                            0.00000000                            0.00000000 
#>             Peptococcus niger et rel.                       Oceanospirillum 
#>                            0.00000000                            0.00000000 
#>                       Novosphingobium                         Moraxellaceae 
#>                            0.00000000                            0.00000000 
#>                        Micrococcaceae                      Methylobacterium 
#>                            0.00000000                            0.00000000 
#>          Megasphaera elsdenii et rel.         Megamonas hypermegale et rel. 
#>                            0.00000000                            0.00000000 
#>                           Leminorella      Lactobacillus salivarius et rel. 
#>                            0.00000000                            0.00000000 
#>       Lactobacillus plantarum et rel.         Lactobacillus gasseri et rel. 
#>                            0.00000000                            0.00000000 
#>    Lactobacillus catenaformis et rel.        Klebisiella pneumoniae et rel. 
#>                            0.00000000                            0.00000000 
#>                          Helicobacter                           Haemophilus 
#>                            0.00000000                            0.00000000 
#>                        Granulicatella                               Gemella 
#>                            0.00000000                            0.00000000 
#>                          Fusobacteria           Eubacterium siraeum et rel. 
#>                            0.00000000                            0.00000000 
#>           Eubacterium limosum et rel.                          Enterococcus 
#>                            0.00000000                            0.00000000 
#>        Enterobacter aerogenes et rel.             Eggerthella lenta et rel. 
#>                            0.00000000                            0.00000000 
#>                 Desulfovibrio et rel.                       Corynebacterium 
#>                            0.00000000                            0.00000000 
#>    Coprobacillus catenaformis et rel.      Clostridium thermocellum et rel. 
#>                            0.00000000                            0.00000000 
#>      Clostridium stercorarium et rel.           Clostridium ramosum et rel. 
#>                            0.00000000                            0.00000000 
#>         Clostridium felsineum et rel.           Clostridium (sensu stricto) 
#>                            0.00000000                            0.00000000 
#>                         Campylobacter                          Burkholderia 
#>                            0.00000000                            0.00000000 
#>              Bulleidia moorei et rel.                           Brachyspira 
#>                            0.00000000                            0.00000000 
#>                     Bilophila et rel.                              Bacillus 
#>                            0.00000000                            0.00000000 
#>                             Atopobium                Asteroleplasma et rel. 
#>                            0.00000000                            0.00000000 
#>                      Aneurinibacillus     Anaerotruncus colihominis et rel. 
#>                            0.00000000                            0.00000000 
#>                          Anaerofustis                    Anaerobiospirillum 
#>                            0.00000000                            0.00000000 
#>          Alcaligenes faecalis et rel.                             Aeromonas 
#>                            0.00000000                            0.00000000 
#>                            Aerococcus                      Actinomycetaceae 
#>                            0.00000000                            0.00000000 

  prevalence(peerj32$phyloseq, detection = 200, sort = TRUE, count = TRUE)

#>            Ruminococcus obeum et rel.        Roseburia intestinalis et rel. 
#>                                    44                                    44 
#>  Faecalibacterium prausnitzii et rel.           Eubacterium rectale et rel. 
#>                                    44                                    44 
#>            Eubacterium hallii et rel.         Dorea formicigenerans et rel. 
#>                                    44                                    44 
#>          Coprococcus eutactus et rel.            Clostridium nexile et rel. 
#>                                    44                                    44 
#>           Anaerostipes caccae et rel.         Ruminococcus lactaris et rel. 
#>                                    44                                    43 
#>         Clostridium symbiosum et rel.        Clostridium sphenoides et rel. 
#>                                    43                                    42 
#>      Bryantella formatexigens et rel.      Subdoligranulum variable at rel. 
#>                                    42                                    41 
#>   Papillibacter cinnamivorans et rel.        Eubacterium ventriosum et rel. 
#>                                    41                                    40 
#>          Bacteroides vulgatus et rel.     Lachnospira pectinoschiza et rel. 
#>                                    40                                    39 
#>          Lachnobacillus bovis et rel.                       Bifidobacterium 
#>                                    39                                    39 
#>           Streptococcus bovis et rel.           Ruminococcus gnavus et rel. 
#>                                    38                                    38 
#>           Ruminococcus bromii et rel.                           Akkermansia 
#>                                    34                                    34 
#>           Streptococcus mitis et rel.         Bacteroides uniformis et rel. 
#>                                    33                                    32 
#>      Bacteroides intestinalis et rel.  Outgrouping clostridium cluster XIVa 
#>                                    31                                    27 
#>            Clostridium leptum et rel.    Oscillospira guillermondii et rel. 
#>                                    26                                    25 
#>      Clostridium orbiscindens et rel.                    Allistipes et rel. 
#>                                    24                                    22 
#>          Bacteroides fragilis et rel.        Oxalobacter formigenes et rel. 
#>                                    17                                    16 
#>    Parabacteroides distasonis et rel.         Clostridium cellulosi et rel. 
#>                                    15                                    15 
#>        Butyrivibrio crossotus et rel.         Ruminococcus callidus et rel. 
#>                                    14                                    13 
#>        Sporobacter termitidis et rel.           Prevotella tannerae et rel. 
#>                                    12                                    11 
#>         Bacteroides stercoris et rel.            Bacteroides ovatus et rel. 
#>                                     9                                     9 
#>            Uncultured Clostridiales I          Bacteroides plebeius et rel. 
#>                                     8                                     8 
#>           Eubacterium biforme et rel.                             Dialister 
#>                                     7                                     7 
#>     Prevotella melaninogenica et rel.              Uncultured Bacteroidetes 
#>                                     6                                     5 
#>     Streptococcus intermedius et rel.                           Collinsella 
#>                                     4                                     4 
#>       Anaerovorax odorimutans et rel.             Prevotella oralis et rel. 
#>                                     4                                     3 
#>                           Lactococcus                              Serratia 
#>                                     3                                     2 
#>     Catenibacterium mitsuokai et rel.              Uncultured Chroococcales 
#>                                     2                                     1 
#>                           Pseudomonas         Prevotella ruminicola et rel. 
#>                                     1                                     1 
#>        Mitsuokella multiacida et rel.      Eubacterium cylindroides et rel. 
#>                                     1                                     1 
#>              Escherichia coli et rel.         Clostridium difficile et rel. 
#>                                     1                                     1 
#>           Clostridium colinum et rel.       Bacteroides splachnicus et rel. 
#>                                     1                                     1 
#>                         Aquabacterium                      Yersinia et rel. 
#>                                     1                                     0 
#>                      Xanthomonadaceae                      Wissella et rel. 
#>                                     0                                     0 
#>                                Vibrio                     Weissella et rel. 
#>                                     0                                     0 
#>                           Veillonella           Uncultured Selenomonadaceae 
#>                                     0                                     0 
#>                 Uncultured Mollicutes           Uncultured Clostridiales II 
#>                                     0                                     0 
#>                    Tannerella et rel.        Sutterella wadsworthia et rel. 
#>                                     0                                     0 
#>                        Staphylococcus                       Proteus et rel. 
#>                                     0                                     0 
#>                     Propionibacterium Phascolarctobacterium faecium et rel. 
#>                                     0                                     0 
#>     Peptostreptococcus micros et rel. Peptostreptococcus anaerobius et rel. 
#>                                     0                                     0 
#>             Peptococcus niger et rel.                       Oceanospirillum 
#>                                     0                                     0 
#>                       Novosphingobium                         Moraxellaceae 
#>                                     0                                     0 
#>                        Micrococcaceae                      Methylobacterium 
#>                                     0                                     0 
#>          Megasphaera elsdenii et rel.         Megamonas hypermegale et rel. 
#>                                     0                                     0 
#>                           Leminorella      Lactobacillus salivarius et rel. 
#>                                     0                                     0 
#>       Lactobacillus plantarum et rel.         Lactobacillus gasseri et rel. 
#>                                     0                                     0 
#>    Lactobacillus catenaformis et rel.        Klebisiella pneumoniae et rel. 
#>                                     0                                     0 
#>                          Helicobacter                           Haemophilus 
#>                                     0                                     0 
#>                        Granulicatella                               Gemella 
#>                                     0                                     0 
#>                          Fusobacteria           Eubacterium siraeum et rel. 
#>                                     0                                     0 
#>           Eubacterium limosum et rel.                          Enterococcus 
#>                                     0                                     0 
#>        Enterobacter aerogenes et rel.             Eggerthella lenta et rel. 
#>                                     0                                     0 
#>                 Desulfovibrio et rel.                       Corynebacterium 
#>                                     0                                     0 
#>    Coprobacillus catenaformis et rel.      Clostridium thermocellum et rel. 
#>                                     0                                     0 
#>      Clostridium stercorarium et rel.           Clostridium ramosum et rel. 
#>                                     0                                     0 
#>         Clostridium felsineum et rel.           Clostridium (sensu stricto) 
#>                                     0                                     0 
#>                         Campylobacter                          Burkholderia 
#>                                     0                                     0 
#>              Bulleidia moorei et rel.                           Brachyspira 
#>                                     0                                     0 
#>                     Bilophila et rel.                              Bacillus 
#>                                     0                                     0 
#>                             Atopobium                Asteroleplasma et rel. 
#>                                     0                                     0 
#>                      Aneurinibacillus     Anaerotruncus colihominis et rel. 
#>                                     0                                     0 
#>                          Anaerofustis                    Anaerobiospirillum 
#>                                     0                                     0 
#>          Alcaligenes faecalis et rel.                             Aeromonas 
#>                                     0                                     0 
#>                            Aerococcus                      Actinomycetaceae 
#>                                     0                                     0