Convert phyloseq-class object to long data format

An alternative to psmelt function from phyloseq-class object.

psmelt2(x, sample.column = NULL, feature.column = NULL)

Arguments

x

phyloseq-class object

sample.column

A single character string specifying name of the column to hold sample names.

feature.column

A single character string specifying name of the column to hold OTU or ASV names.

Value

A tibble in long format

Author

Contact: Sudarshan A. Shetty sudarshanshetty9@gmail.com

Examples

data("dietswap")
ps.melt <- psmelt2(dietswap, sample.column="SampleID", 
                   feature.column="Feature") 
head(ps.melt)                                         
#> # A tibble: 6 × 14
#>   Feature     SampleID value Phylum Family Genus subject sex   nationality group
#>   <chr>       <chr>    <dbl> <chr>  <chr>  <chr> <fct>   <fct> <fct>       <fct>
#> 1 Actinomyce… Sample-1     0 Actin… Actin… Acti… byn     male  AAM         DI   
#> 2 Actinomyce… Sample-2     1 Actin… Actin… Acti… nms     male  AFR         HE   
#> 3 Actinomyce… Sample-3     0 Actin… Actin… Acti… olt     male  AFR         HE   
#> 4 Actinomyce… Sample-4     1 Actin… Actin… Acti… pku     fema… AFR         HE   
#> 5 Actinomyce… Sample-5     0 Actin… Actin… Acti… qjy     fema… AFR         HE   
#> 6 Actinomyce… Sample-6     0 Actin… Actin… Acti… riv     fema… AFR         HE   
#> # ℹ 4 more variables: sample <chr>, timepoint <int>,
#> #   timepoint.within.group <int>, bmi_group <fct>