Read Simple OTU Tables into a Phyloseq Object

Read simple OTU tables, mapping and taxonomy files into a phyloseq-class object.

read_csv2phyloseq(
  otu.file = NULL,
  taxonomy.file = NULL,
  metadata.file = NULL,
  sep = ","
)

Arguments

otu.file

A simple otu_table with '.csv' extension

taxonomy.file

A simple taxonomy file with '.csv' extension

metadata.file

A simple metadata/mapping file with .csv extension

sep

CSV file separator

Value

phyloseq-class object.

Details

Simple OTU tables, mapping and taxonomy files will be converted to phyloseq-class.

Author

Sudarshan A. Shetty sudarshanshetty9@gmail.com

Examples

# NOTE: the system.file command reads these example files from the
# microbiome R package. To use your own local files, simply write
# otu.file <- "/path/to/my/file.csv" etc.

#otu.file <-
#   system.file("extdata/qiita1629_otu_table.csv",
#   package='microbiome')

#tax.file <- system.file("extdata/qiita1629_taxonomy_table.csv",
#        package='microbiome')

#meta.file <- system.file("extdata/qiita1629_mapping_subset.csv",
#     package='microbiome')

#p0 <- read_csv2phyloseq(
#        otu.file=otu.file, 
#        taxonomy.file=tax.file, 
#        metadata.file=meta.file)