Taxa Names

List the names of taxonomic groups in a phyloseq object.

taxa(x)

Arguments

x

phyloseq-class object

Value

A vector with taxon names.

Details

A handy shortcut for phyloseq::taxa_names, with a potential to add to add some extra tweaks later.

References

To cite the microbiome R package, see citation('microbiome')

Examples

  data(dietswap)
  taxa(dietswap)

#>   [1] "Actinomycetaceae"                     
#>   [2] "Aerococcus"                           
#>   [3] "Aeromonas"                            
#>   [4] "Akkermansia"                          
#>   [5] "Alcaligenes faecalis et rel."         
#>   [6] "Allistipes et rel."                   
#>   [7] "Anaerobiospirillum"                   
#>   [8] "Anaerofustis"                         
#>   [9] "Anaerostipes caccae et rel."          
#>  [10] "Anaerotruncus colihominis et rel."    
#>  [11] "Anaerovorax odorimutans et rel."      
#>  [12] "Aneurinibacillus"                     
#>  [13] "Aquabacterium"                        
#>  [14] "Asteroleplasma et rel."               
#>  [15] "Atopobium"                            
#>  [16] "Bacillus"                             
#>  [17] "Bacteroides fragilis et rel."         
#>  [18] "Bacteroides intestinalis et rel."     
#>  [19] "Bacteroides ovatus et rel."           
#>  [20] "Bacteroides plebeius et rel."         
#>  [21] "Bacteroides splachnicus et rel."      
#>  [22] "Bacteroides stercoris et rel."        
#>  [23] "Bacteroides uniformis et rel."        
#>  [24] "Bacteroides vulgatus et rel."         
#>  [25] "Bifidobacterium"                      
#>  [26] "Bilophila et rel."                    
#>  [27] "Brachyspira"                          
#>  [28] "Bryantella formatexigens et rel."     
#>  [29] "Bulleidia moorei et rel."             
#>  [30] "Burkholderia"                         
#>  [31] "Butyrivibrio crossotus et rel."       
#>  [32] "Campylobacter"                        
#>  [33] "Catenibacterium mitsuokai et rel."    
#>  [34] "Clostridium (sensu stricto)"          
#>  [35] "Clostridium cellulosi et rel."        
#>  [36] "Clostridium colinum et rel."          
#>  [37] "Clostridium difficile et rel."        
#>  [38] "Clostridium felsineum et rel."        
#>  [39] "Clostridium leptum et rel."           
#>  [40] "Clostridium nexile et rel."           
#>  [41] "Clostridium orbiscindens et rel."     
#>  [42] "Clostridium ramosum et rel."          
#>  [43] "Clostridium sphenoides et rel."       
#>  [44] "Clostridium stercorarium et rel."     
#>  [45] "Clostridium symbiosum et rel."        
#>  [46] "Clostridium thermocellum et rel."     
#>  [47] "Collinsella"                          
#>  [48] "Coprobacillus catenaformis et rel."   
#>  [49] "Coprococcus eutactus et rel."         
#>  [50] "Corynebacterium"                      
#>  [51] "Desulfovibrio et rel."                
#>  [52] "Dialister"                            
#>  [53] "Dorea formicigenerans et rel."        
#>  [54] "Eggerthella lenta et rel."            
#>  [55] "Enterobacter aerogenes et rel."       
#>  [56] "Enterococcus"                         
#>  [57] "Escherichia coli et rel."             
#>  [58] "Eubacterium biforme et rel."          
#>  [59] "Eubacterium cylindroides et rel."     
#>  [60] "Eubacterium hallii et rel."           
#>  [61] "Eubacterium limosum et rel."          
#>  [62] "Eubacterium rectale et rel."          
#>  [63] "Eubacterium siraeum et rel."          
#>  [64] "Eubacterium ventriosum et rel."       
#>  [65] "Faecalibacterium prausnitzii et rel." 
#>  [66] "Fusobacteria"                         
#>  [67] "Gemella"                              
#>  [68] "Granulicatella"                       
#>  [69] "Haemophilus"                          
#>  [70] "Helicobacter"                         
#>  [71] "Klebisiella pneumoniae et rel."       
#>  [72] "Lachnobacillus bovis et rel."         
#>  [73] "Lachnospira pectinoschiza et rel."    
#>  [74] "Lactobacillus catenaformis et rel."   
#>  [75] "Lactobacillus gasseri et rel."        
#>  [76] "Lactobacillus plantarum et rel."      
#>  [77] "Lactobacillus salivarius et rel."     
#>  [78] "Lactococcus"                          
#>  [79] "Leminorella"                          
#>  [80] "Megamonas hypermegale et rel."        
#>  [81] "Megasphaera elsdenii et rel."         
#>  [82] "Methylobacterium"                     
#>  [83] "Micrococcaceae"                       
#>  [84] "Mitsuokella multiacida et rel."       
#>  [85] "Moraxellaceae"                        
#>  [86] "Novosphingobium"                      
#>  [87] "Oceanospirillum"                      
#>  [88] "Oscillospira guillermondii et rel."   
#>  [89] "Outgrouping clostridium cluster XIVa" 
#>  [90] "Oxalobacter formigenes et rel."       
#>  [91] "Papillibacter cinnamivorans et rel."  
#>  [92] "Parabacteroides distasonis et rel."   
#>  [93] "Peptococcus niger et rel."            
#>  [94] "Peptostreptococcus anaerobius et rel."
#>  [95] "Peptostreptococcus micros et rel."    
#>  [96] "Phascolarctobacterium faecium et rel."
#>  [97] "Prevotella melaninogenica et rel."    
#>  [98] "Prevotella oralis et rel."            
#>  [99] "Prevotella ruminicola et rel."        
#> [100] "Prevotella tannerae et rel."          
#> [101] "Propionibacterium"                    
#> [102] "Proteus et rel."                      
#> [103] "Pseudomonas"                          
#> [104] "Roseburia intestinalis et rel."       
#> [105] "Ruminococcus bromii et rel."          
#> [106] "Ruminococcus callidus et rel."        
#> [107] "Ruminococcus gnavus et rel."          
#> [108] "Ruminococcus lactaris et rel."        
#> [109] "Ruminococcus obeum et rel."           
#> [110] "Serratia"                             
#> [111] "Sporobacter termitidis et rel."       
#> [112] "Staphylococcus"                       
#> [113] "Streptococcus bovis et rel."          
#> [114] "Streptococcus intermedius et rel."    
#> [115] "Streptococcus mitis et rel."          
#> [116] "Subdoligranulum variable at rel."     
#> [117] "Sutterella wadsworthia et rel."       
#> [118] "Tannerella et rel."                   
#> [119] "Uncultured Bacteroidetes"             
#> [120] "Uncultured Chroococcales"             
#> [121] "Uncultured Clostridiales I"           
#> [122] "Uncultured Clostridiales II"          
#> [123] "Uncultured Mollicutes"                
#> [124] "Uncultured Selenomonadaceae"          
#> [125] "Veillonella"                          
#> [126] "Vibrio"                               
#> [127] "Weissella et rel."                    
#> [128] "Wissella et rel."                     
#> [129] "Xanthomonadaceae"                     
#> [130] "Yersinia et rel."