Variation Line Plot

Plot variation in taxon abundance for many subjects.

tipplot(x, taxon, tipping.point = NULL, lims = NULL, shift = 0.001,
  xlim = NULL)

Arguments

x

phyloseq-class object

taxon

Taxonomic group to visualize.

tipping.point

Optional. Indicate critical point for abundance variations to be highlighted.

lims

Optional. Figure X axis limits.

shift

Small constant to avoid problems with zeroes in log10

xlim

Horizontal axis limits

Value

ggplot object

Details

Assuming the sample_data(x) has 'subject' field and some subjects have multiple time points.

References

See citation('microbiome')

Examples

  data(atlas1006)
  pseq <- atlas1006
  pseq <- subset_samples(pseq, DNA_extraction_method == "r")
  pseq <- transform(pseq, "compositional")
  p <- tipplot(pseq, "Dialister", tipping.point = 1)