Diversity Index

Various community diversity indices.

diversity(x, index = "all", zeroes = TRUE)

Arguments

x

A species abundance vector, or matrix (taxa/features x samples) with the absolute count data (no relative abundances), or phyloseq-class object

index

Diversity index. See details for options.

zeroes

Include zero counts in the diversity estimation.

Value

A vector of diversity indices

Details

By default, returns all diversity indices. The available diversity indices include the following:

inverse_simpson Inverse Simpson diversity: $1/lambda$ where $lambda=sum(p^2)$ and $p$ are relative abundances.

gini_simpson Gini-Simpson diversity $1 - lambda$. This is also called Gibbs–Martin, or Blau index in sociology, psychology and management studies.

shannon Shannon diversity ie entropy

fisher Fisher alpha; as implemented in the vegan package

coverage Number of species needed to cover 50% of the ecosystem. For other quantiles, apply the function coverage directly.

References

Beisel J-N. et al. A Comparative Analysis of Diversity Index Sensitivity. Internal Rev. Hydrobiol. 88(1):3-15, 2003. URL: https://portais.ufg.br/up/202/o/2003-comparative_evennes_index.pdf

Bulla L. An index of diversity and its associated diversity measure. Oikos 70:167--171, 1994

Magurran AE, McGill BJ, eds (2011) Biological Diversity: Frontiers in Measurement and Assessment (Oxford Univ Press, Oxford), Vol 12.

Smith B and Wilson JB. A Consumer's Guide to Diversity Indices. Oikos 76(1):70-82, 1996.

See also

Author

Contact: Leo Lahti microbiome-admin@googlegroups.com

Examples

data(dietswap)
d <- alpha(dietswap, 'shannon')
#> Observed richness
#> Other forms of richness
#> Diversity
#> Evenness
#> Dominance
#> Rarity