Visualize OTU Core

Core visualization (2D).

plot_core(x, prevalences = seq(, 1, 1, 0.1), detections = 20,
  plot.type = "lineplot", colours = gray(seq(0, 1, length = 5)),
  min.prevalence = NULL, taxa.order = NULL, horizontal = FALSE)

Arguments

x

A phyloseq object or a core matrix

prevalences

a vector of prevalence percentages in [0,1]

detections

a vector of intensities around the data range, or a scalar indicating the number of intervals in the data range.

plot.type

Plot type ('lineplot' or 'heatmap')

colours

colours for the heatmap

min.prevalence

If minimum prevalence is set, then filter out those rows (taxa) and columns (detections) that never exceed this prevalence. This helps to zoom in on the actual core region of the heatmap. Only affects the plot.type = 'heatmap'.

taxa.order

Ordering of the taxa.

horizontal

Logical. Horizontal figure.

Value

A list with three elements: the ggplot object and the data. The data has a different form for the lineplot and heatmap. Finally, the applied parameters are returned.

References

A Salonen et al. The adult intestinal core microbiota is determined by analysis depth and health status. Clinical Microbiology and Infection 18(S4):16 20, 2012. To cite the microbiome R package, see citation('microbiome')

Examples

  data(atlas1006)
  p <- plot_core(atlas1006, prevalences = seq(0.1, 1, .1),
                       detections = c(0, 10^(0:4)))